Fix requested changes
Refactored api controllers and routes
