+++ .gitignore
@@ -3,3 +3,4 @@ ml/data/face_inference.mp4
 ml/data/test_images
 node_modules
 ui/out
+__debug_bin*
