+++ api/controllers/locations.go
@@ -1,6 +1,9 @@
 package controllers
 
 import (
+	"strconv"
+
+	"github.com/gin-gonic/gin"
 	"github.com/uug-ai/facial-access-control/api/database"
 	"github.com/uug-ai/facial-access-control/api/models"
 )
@@ -12,9 +15,13 @@ import (
 // @Summary Get all locations
 // @Description Get all locations
 // @Success 200 {object} []models.Location
-func GetLocations() []models.Location {
-	locations := database.GetLocations()
-	return locations
+func GetLocations(c *gin.Context) []models.Location {
+	// Create a list of random locations
+			locations := database.GetLocations()
+			c.JSON(200, gin.H{
+				"data": locations,
+			})
+			return nil;
 }
 
 // Location godoc
@@ -25,14 +32,26 @@ func GetLocations() []models.Location {
 // @Description Get location by ID
 // @Param id path int true "Location ID"
 // @Success 200 {object} models.Location
-func GetLocation(id int) models.Location {
-	locations := database.GetLocations()
-	for _, location := range locations {
-		if location.Id == id {
+func GetLocation(c *gin.Context) models.Location {
+				// Get the id parameter from the URL
+			id := c.Param("id")
+
+			// Convert id to an integer
+			locationID, err := strconv.Atoi(id)
+			if err != nil {
+				c.JSON(400, gin.H{
+					"error": "Invalid location ID",
+				})
+				return models.Location{}
+			}
+
+			// Use the locationID to fetch the location
+			location := database.GetLocation(locationID)
+
+			c.JSON(200, gin.H{
+				"data": location,
+			})
 			return location
-		}
-	}
-	return models.Location{}
 }
 
 
@@ -46,9 +65,27 @@ func GetLocation(id int) models.Location {
 // @Produce json
 // @Param location body models.Location true "Location"
 // @Success 200 {object} models.Location
-func AddLocation(location models.Location) error {
-	err := database.AddLocation(location)
+func AddLocation(c *gin.Context) error {
+	var location models.Location
+	if err := c.ShouldBindJSON(&location); err != nil {
+		c.JSON(400, gin.H{
+			"error": "Invalid location data",
+		})
+		return err
+	}
+
+	if err := database.AddLocation(location); err != nil {
+		c.JSON(500, gin.H{
+			"error": "Failed to add location",
+		})
 	return err
+	}
+
+	c.JSON(201, gin.H{
+		"message": "Location added successfully",
+		"location": location,
+	})
+	return nil	
 }
 
 // location godoc
@@ -59,9 +96,28 @@ func AddLocation(location models.Location) error {
 // @Description Delete location
 // @Param id path int true "Location ID"
 // @Success 200
-func DeleteLocation(id int) error {
-	err := database.DeleteLocation(id)
-	return err
+func DeleteLocation(c *gin.Context) error {
+	id := c.Param("id")
+	locationID, err := strconv.Atoi(id)
+	
+	if err != nil {
+		c.JSON(400, gin.H{
+			"error": "Invalid location ID",
+		})
+		return err
+	}
+
+	if err := database.DeleteLocation(locationID); err != nil {
+		c.JSON(500, gin.H{
+			"error": "Failed to delete location",
+		})
+		return err
+	}
+
+	c.JSON(200, gin.H{
+		"message": "Location deleted successfully",
+	})
+	return nil
 }
